refactor(cli): dedupe validate report ownership
Move report models and formatting into the validate report submodule and keep the command module focused on orchestration.
